fix tensor slice worker distribution bug
